Remove old blobs and hex edit i9305/proprietary/system/lib/libsec-ril.so
This is an old blob from a system which only had mediaserver. In Nougat
audioserver has been created. So you need to search for the string:

        /system/bin/mediaserver

You need to replace 'mediaserver' with 'audioserver'. Right before this
string you will find the string 'media' which is group the process
is running as, replace that with 'audio'.